TEXTTAB workaround
Former-commit-id: 249f10b0dff61458cd248c7af98190af9a4940ba
Former-commit-id: 805255371ef1d1abd68d4cf0b850a333c2660c91

+++ src/engine/rendertext.cpp
@@ -128,7 +128,7 @@ float text_widthf(const char *str, int flags)
     return width;
 }
 
-#define TEXTTAB(g) ((g) + max(float(FONTTAB - fmod((g), FONTTAB)), float(FONTW)))
+#define TEXTTAB(g) ((g) + max(float(FONTTAB - fmodf((g), FONTTAB)), float(FONTW)))
 
 void tabify(const char *str, int *numtabs)
 {
